한 사이클로 돌아가게 만들었음. 그래도 버그난거 다시 고쳐야함....

+++ yaCharacterIcon.cpp
@@ -42,14 +42,14 @@ namespace ya
 				hdc,
 				static_cast<int>(mScreenPos.x),
 				static_cast<int>(mScreenPos.y),
-				static_cast<int>(mNoSelelctedImgWidth),
-				static_cast<int>(mNoSelelctedImgHeight),
+				mNoSelelctedImgWidth,
+				mNoSelelctedImgHeight,
 			
 				mpCurrRenderImage->GetDC(),
 				0,
 				0,
-				static_cast<int>(mNoSelelctedImgWidth),
-				static_cast<int>(mNoSelelctedImgHeight),
+				mNoSelelctedImgWidth,
+				mNoSelelctedImgHeight,
 				mBlendFunc
 			);
 		}
@@ -59,14 +59,14 @@ namespace ya
 				hdc,
 				static_cast<int>(mScreenPos.x + mOffset.x),
 				static_cast<int>(mScreenPos.y + mOffset.y),
-				static_cast<int>(mSelelctedImgWidth),
-				static_cast<int>(mSelelctedImgHeight),
+				mSelelctedImgWidth,
+				mSelelctedImgHeight,
 
 				mpCurrRenderImage->GetDC(),
 				0,
 				0,
-				static_cast<int>(mSelelctedImgWidth),
-				static_cast<int>(mSelelctedImgHeight),
+				mSelelctedImgWidth,
+				mSelelctedImgHeight,
 				mBlendFunc
 			);
 		}
